+++ src/main/java/Command.java
@@ -1,3 +1,5 @@
+package src.main.java;
+
 public interface Command {
     Lexeme execute();
 }
