+++ src/main/java/Multiply.java
@@ -1,13 +1,15 @@
-public class Sub implements Command{
+package src.main.java;
+
+public class Multiply implements Command {
     private Numbers numbers;
 
-    public Sub(Numbers numbers) {
+    public Multiply(Numbers numbers) {
         this.numbers = numbers;
     }
 
     @Override
     public Lexeme execute() {
-        String value = String.valueOf(numbers.numb2 - numbers.numb1);
+        String value = String.valueOf(numbers.numb1 * numbers.numb2);
         Lexeme lexeme = new Lexeme(TypeLexeme.NUMBER, 0, value);
         return lexeme;
     }
