+++ src/main/java/StackHolder.java
@@ -1,3 +1,5 @@
+package src.main.java;
+
 import java.util.Stack;
 
 public class StackHolder {
