+++ src/main/java/Sub.java
@@ -1,13 +1,15 @@
-public class Multiply implements Command{
+package src.main.java;
+
+public class Sub implements Command {
     private Numbers numbers;
 
-    public Multiply(Numbers numbers) {
+    public Sub(Numbers numbers) {
         this.numbers = numbers;
     }
 
     @Override
     public Lexeme execute() {
-        String value = String.valueOf(numbers.numb1 * numbers.numb2);
+        String value = String.valueOf(numbers.numb2 - numbers.numb1);
         Lexeme lexeme = new Lexeme(TypeLexeme.NUMBER, 0, value);
         return lexeme;
     }
