+++ src/main/java/TypeBracket.java
@@ -1,3 +1,5 @@
+package src.main.java;
+
 public enum TypeBracket {
     OPEN,
     CLOSE
