+++ src/main/java/TypeLexeme.java
@@ -1,3 +1,5 @@
+package src.main.java;
+
 public enum TypeLexeme {
     COMMAND,
     NUMBER,
